+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.21 - 26 August 2026
 1. Make new script executable
+2. Refactor root.py into main.py
 
 ## Version 1.0.20 - 25 August 2026
 1. Add ability to open project in Dolphin