+++ HISTORY.md
@@ -4,6 +4,7 @@
 1. Make new script executable
 2. Refactor root.py into main.py
 3. Change call params to BuildFrame to pass self instead of self.root
+4. Refactor module_caller.py for better error handling and validation
 
 ## Version 1.0.20 - 25 August 2026
 1. Add ability to open project in Dolphin